# docs/api/swagger.yaml
definitions:
...
  payload.HTTPMessage:
    description: HTTPMessage represents a message to be sent to the client trough
      HTTP REST API.
    properties:
      code:
        description: |-
          Code is a stable, machine-readable reason, present only where a client
          has to BRANCH on it. Message is prose and may be reworded; this is not.

          It exists because a 401 has two meanings a client must treat differently:
          an expired access token should be refreshed and the request retried,
          while a revoked one must not be — the refresh token was revoked in the
          same breath, so the retry burns two requests and fails anyway. Nothing in
          the old response distinguished them, and a client cannot be asked to
          match on English.

          Empty on every response that does not need one, so no existing client
          sees a change.
        example: token_revoked
        format: string
        type: string
      message:
        example: success
        format: string
        type: string
      method:
        example: GET
        format: string
        type: string
      path:
        example: /api/v1/users
        format: string
        type: string
      status_code:
        example: 200
        format: int32
        type: integer
      timestamp:
        example: "2021-07-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  payload.IDPLoginResponse:
    description: Response containing OAuth redirect information for initiating login
      flow
    properties:
      idp_id:
        example: 019b4b0d-a682-7e25-b4c8-afa26aa7d1dc
        format: uuid
        type: string
      redirect_code:
        example: 302
        type: integer
      redirect_url:
        format: uri
        type: string
    type: object
  payload.IDPRegisterResponse:
    description: Response containing OAuth redirect information for initiating registration
      flow
    properties:
      idp_id:
        example: 019b4b0d-a682-7e29-b6e2-a4cc6db87ea8
        format: uuid
        type: string
      redirect_code:
        example: 302
        type: integer
      redirect_url:
        format: uri
        type: string
    type: object
...
  payload.ListIDPAvailableResponse:
    description: List of identity providers available for configuration in the system
    properties:
      items:
        description: Array of available IDP options
        items:
          $ref: '#/definitions/payload.IDPAvailableResponse'
        type: array
    type: object
...
  payload.ListIDPsResponse:
    description: Paginated list of identity provider configurations
    properties:
      items:
        description: Array of IDP configurations
        items:
          $ref: '#/definitions/payload.IDPResponse'
        type: array
      paginator:
        allOf:
        - $ref: '#/definitions/domain.Paginator'
        description: Pagination information
    type: object
...
info:
  contact:
    email: info@goapitemplate.local
    name: API Support
    url: https://goapitemplate.local
  description: |-
    A production-shaped starting point for a Go HTTP REST API: hexagonal architecture, multi-tenant projects, RBAC through Open Policy Agent, database-backed rate limiting and resource limits, and OpenTelemetry traces and metrics throughout.

    `products` is the worked example entity -- a project-scoped CRUD resource that exercises every convention here. Copy it when adding your own; see docs/architecture/adding-an-entity.md.

    ### Authentication

    Four bearer credentials, each accepted by a different set of endpoints; see the security definitions below for which is which. Ordinary calls use **AccessToken**.

    A logout revokes the access token it was called with, not only the refresh token, so it stops working immediately. That check can be switched off (`authn.access.token.revocation.enabled`), and where it is, a logged-out access token keeps working until it expires -- which is why the lifetime is short.

    Refresh tokens are rotated: each refresh spends the token it was given and returns a successor, so a client **must** store what `POST /auth/refresh` returns. Presenting a spent one again reads as a replay and ends the session for whoever holds the live token too.

    ### Conventions

    - **Identifiers are UUID v7.** Every `id` in this API, without exception.
...

    ### Rate limiting

...

...

...

    ### Operations

...
  version: v1
paths:
  /auth/idp/{idp_id}/callback:
    get:
      consumes:
      - application/json
      description: Process OAuth callback from Identity Provider, validates state
        and authorization code.
      operationId: 01988e60-89e5-72ee-9db4-db5cd7535717
      parameters:
      - description: Identity Provider unique identifier
        format: uuid
        in: path
        name: idp_id
        required: true
        type: string
      - description: OAuth state parameter for CSRF protection
        in: query
        name: state
        required: true
        type: string
      - description: OAuth authorization code from IDP
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Callback processed successfully. Authentication cookies are
            set and the caller is redirected to the IDP's configured login or register
            redirect URL. This endpoint never returns a JSON body on success — it
            always issues a redirect.
          headers:
            Location:
              description: The IDP's configured LoginRedirectURL or RegisterRedirectURL
              type: string
          schema:
            type: string
        "400":
          description: Invalid parameters, missing state/code, or invalid IDP ID format
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "404":
          description: Identity Provider not found
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "409":
          description: User already exists during registration
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error during callback processing
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      summary: Handle IDP OAuth callback
      tags:
      - Authentication
  /auth/idp/{idp_id}/login:
    get:
      consumes:
      - application/json
      description: Initiate authentication with specified Identity Provider and returns
        redirect URL for OAuth flow.
      operationId: 01988e60-89e5-72ab-adb4-3eef95d1afd3
      parameters:
      - description: Identity Provider unique identifier
        format: uuid
        in: path
        name: idp_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login URL generated successfully. RedirectURL and RedirectCode
            are fields of the JSON body — this endpoint returns 200, it does not itself
            redirect.
          schema:
            $ref: '#/definitions/payload.IDPLoginResponse'
        "400":
          description: Invalid IDP ID format or malformed request
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: 'Internal server error during URL generation. NOTE: an unknown
            Identity Provider currently surfaces here rather than as a 404 — the handler
            has no not-found branch. Tracked as a behaviour defect; this annotation
            documents what the endpoint does today, not what it should do.'
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      summary: Initiate IDP login
      tags:
      - Authentication
  /auth/idp/{idp_id}/register:
    get:
      consumes:
      - application/json
      description: Initiate user registration with specified Identity Provider and
        returns redirect URL for OAuth registration flow.
      operationId: 019894ba-6014-79cf-bff4-6668484cc7e3
      parameters:
      - description: Identity Provider unique identifier
        format: uuid
        in: path
        name: idp_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Registration URL generated successfully. RedirectURL and RedirectCode
            are fields of the JSON body — this endpoint returns 200, it does not itself
            redirect.
          schema:
            $ref: '#/definitions/payload.IDPRegisterResponse'
        "400":
          description: Invalid IDP ID format or malformed request
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: 'Internal server error during URL generation. NOTE: an unknown
            Identity Provider currently surfaces here rather than as a 404 — the handler
            has no not-found branch. Tracked as a behaviour defect; this annotation
            documents what the endpoint does today, not what it should do.'
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      summary: Initiate IDP registration
      tags:
      - Authentication
  /auth/idp/available:
    get:
      consumes:
      - application/json
      description: Retrieve all identity providers configured and available for user
        authentication and registration.
      operationId: 0198fb33-7333-76f9-bcb4-1af086de3e10
      produces:
      - application/json
      responses:
        "200":
          description: List of available Identity Providers retrieved successfully
          schema:
            $ref: '#/definitions/payload.ListIDPAvailableResponse'
        "400":
          description: Malformed request
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error retrieving IDPs
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      summary: List identity providers
      tags:
      - Authentication
...
  /auth/idps:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of Identity Providers with optional filtering
        and sorting
      operationId: 0198e7ea-3755-7a29-90ed-13245b54f074
      parameters:
      - description: Sort order (e.g., name ASC, created_at DESC)
        in: query
        name: sort
        type: string
      - description: Filter expression (e.g., idp_type='oauth2' AND name LIKE 'Google%')
        in: query
        name: filter
        type: string
      - description: Comma-separated fields to return (e.g., id,name,idp_type)
        in: query
        name: fields
        type: string
      - description: Pagination token for next page
        in: query
        name: next_token
        type: string
      - description: Pagination token for previous page
        in: query
        name: prev_token
        type: string
      - description: 'Maximum number of items per page (default: system-defined)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of IDPs with pagination metadata
          schema:
            $ref: '#/definitions/payload.ListIDPsResponse'
        "400":
          description: Invalid query parameters, filter syntax, or sort fields
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: List IDPs
      tags:
      - Authentication
      - Identity Providers (IDPs)
    post:
      consumes:
      - application/json
      description: Register a new Identity Provider with authentication configuration
      operationId: 0198e7ea-3755-7a39-9dfc-717d83facf02
      parameters:
      - description: IDP configuration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.CreateIDPRequest'
      produces:
      - application/json
      responses:
        "201":
          description: IDP created successfully
          headers:
            Location:
              description: URL of the newly created IDP resource (/auth/idps/{idp_id})
              type: string
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "400":
          description: Invalid request body, validation failure, or malformed UUID
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "401":
          description: Missing or invalid authentication token
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "409":
          description: IDP with this name or configuration already exists
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "429":
          description: Too many requests -- RATE_LIMIT_EXCEEDED is the budget, RATE_LIMIT_UNAVAILABLE
            the limiter's own store
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/payload.HTTPMessage'
      security:
      - AccessToken: []
      summary: Create IDP
      tags:
      - Authentication
      - Identity Providers (IDPs)
...
securityDefinitions:
  AccessToken:
    description: 'The ordinary credential: `Bearer <access_token>`, from `POST /auth/login`
      or `POST /auth/refresh`. Every endpoint outside `/auth`, `/health` and `/version`
      takes it. A personal access token is presented the same way and is accepted
      anywhere an access token is -- which is how a script authenticates without a
      login. Deliberately short-lived; refresh rather than lengthening it.'
    in: header
    name: Authorization
    type: apiKey
  RefreshToken:
    description: '`Bearer <refresh_token>`, and accepted **only** by `POST /auth/refresh`
      and `DELETE /auth/logout`. Refreshing spends it and returns a successor: store
      what comes back. Presenting a spent one is treated as a replay and ends the
      whole session, including for whoever holds the live token.'
    in: header
    name: Authorization
    type: apiKey
  ResetPasswordToken:
    description: '`Bearer <reset_password_token>`, accepted only by `POST /auth/password/reset`.
      Issued by `POST /auth/password/recover` and delivered by email, single use.'
    in: header
    name: Authorization
    type: apiKey
  VerificationToken:
    description: '`Bearer <verification_token>`, accepted only by `POST /auth/verify/confirm`.
      It proves an email address rather than exercising a permission, so it is the
      one credential no authorization check is applied to -- the account it belongs
      to is still disabled at that point. It never travels in a URL: the emailed link
      points at the frontend, which presents the token in this header.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...